+++ libraries/matchers.rb
@@ -11,15 +11,17 @@
 #     expect(chef_run).to install_chocolatey('cool_package')
 #
 if defined?(ChefSpec)
+  ChefSpec.define_matcher(:chocolatey)
+
   def install_chocolatey(name)
-    ChefSpec::Matchers::ResourceMatcher.new('chocolatey', :install, name)
+    ChefSpec::Matchers::ResourceMatcher.new(:chocolatey, :install, name)
   end
 
   def upgrade_chocolatey(name)
-    ChefSpec::Matchers::ResourceMatcher.new('chocolatey', :upgrade, name)
+    ChefSpec::Matchers::ResourceMatcher.new(:chocolatey, :upgrade, name)
   end
 
   def remove_chocolatey(name)
-    ChefSpec::Matchers::ResourceMatcher.new('chocolatey', :remove, name)
+    ChefSpec::Matchers::ResourceMatcher.new(:chocolatey, :remove, name)
   end
 end